Completely revamped the project
After a longer break, I've decided to completely revamp the project.

The overall project will now be able to cover more smaller projects.

Regarding the hotswap thingy i was working on, I'm taking a slightly different approach, and decided to aim for general GPUs, instead of specifically hybrid graphics system.
If succesfull, it'll work the same on a hybrid system, as the initial approach, but this newer approach should allow for more compatibility with desktops, servers, etc.